test: move double-done coverage into done-async.test.ts and run it concurrently
